+++ web_customizer/tests/__init__.py
@@ -5,6 +5,7 @@ from . import test_js
 from . import test_ledger
 from . import test_menu
 from . import test_optional_column
+from . import test_search_filter
 from . import test_security
 from . import test_tour
 from . import test_ui